+++ R/gsSurvCalendar.R
@@ -163,6 +163,9 @@ gsSurvCalendar <- function(
   y$T <- calendarTime
   y$method <- x$method
   y$call <- match.call()
+  input_vals$testUpper <- testUpper
+  if (test.type > 2) input_vals$testLower <- testLower
+  if (test.type > 6) input_vals$testHarm <- testHarm
   y$inputs <- input_vals
   class(y) <- c("gsSurv", "gsDesign")